jeff


git-svn-id: https://svn.internet2.edu/svn/owamp/trunk@660 a2d82226-602c-0410-bba7-e51b83dd10d2

+++ owping/owpingP.h
@@ -43,7 +43,7 @@ typedef	struct {
 		char		*identity;        /* -u */
 		u_int32_t	numPackets;       /* -c */
 
-		u_int32_t	lossThreshold;    /* -l */
+		double		lossThreshold;    /* -l */
 		float           percentile;       /* -a */
 
 		char		*passwd;          /* -P */
